Collapsed some Plot styling helper structs in to a single file
