--- Sources/SwiftPlot/PlotBorder.swift
@@ -1,5 +1,0 @@
-public struct PlotBorder{
-    public var color = Color.black
-    public var thickness: Float = 2
-    public init() {}
-}
