--- Sources/SwiftPlot/PlotLegend.swift
@@ -1,8 +1,0 @@
-public struct PlotLegend {
-    public var backgroundColor = Color.transluscentWhite
-    public var borderColor = Color.black
-    public var borderThickness: Float = 2
-    public var textColor = Color.black
-    public var textSize: Float = 10
-    public init() {}
-}
