--- Sources/SwiftPlot/PlotTitle.swift
@@ -1,8 +1,0 @@
-public struct PlotTitle {
-    public var title = "TITLE"
-    public var color = Color.black
-    public var size: Float = 20
-    public init(_ title: String = "TITLE") {
-      self.title = title
-    }
-}
